Moved uploadPhoto method to new class PhotoNetworkService
