+++ MultipartRequestCamera/Network/PagesNetworkService.swift
@@ -8,7 +8,6 @@
 import Foundation
 
 final class PagesNetworkService {
-    private let uploadPhotoPath:String = "/api/v2/photo"
     private let getPageTypePath: String = "/api/v2/photo/type"
     private var baseURLComponents: URLComponents {
         var components = URLComponents()
@@ -40,28 +39,5 @@ final class PagesNetworkService {
             task.resume()
         }
     
-    func uploadPhoto(
-        pageContent: PageContent,
-        fileName:String,
-        photo: Data,
-        fileType:MultipartRequest.FileType = .jpeg,
-        complition: @escaping (Error?) -> ()) {
-            var photoUploadComponents = baseURLComponents
-            photoUploadComponents.path = uploadPhotoPath
-            
-            var request = MultipartRequest(url: photoUploadComponents.url!)
-            request.addFormField(
-                fieldName: "name",
-                value: pageContent.name)
-            request.addFile(
-                fieldName: "photo",
-                fileName: fileName,
-                file: photo,
-                fileType: fileType)
-            request.addFormField(
-                fieldName: "typeId",
-                value: String(pageContent.id))
-            
-            request.sendRequest(complition: complition)
-        }
+    
 }
